Refactor clustering to be "component first" instead of "contiBAIT first"
There is now a very strong "each component comes from one chromosome assumption" (with the exception of the acrocentrics.). Previously, the contiBAIT clustering was the central method, with some code to apply that assumption if it appeared to be confirmed by the contiBAIT clustering results. However, for the data that the pipeline has been developed on, the contiBAIT clustering is not-uncommonly worse than the 1-component 1-chromosome assumption. Even with very high-quality graphs, the contiBait clustering often bisects chromosomes. In general, the contiBAIT clustering results exhibit too much variance. Now the 1-component, 1-chromosome assumption is central, and the contiBAIT clustering supports it by linking stray unitigs to larger clusters.