reduce cores for reference mapping
From peter: I just noticed a bunch of your jobs waiting in the HiMem queue on HILBERT, despite a resource spec of only ~98 GB. I presume the ncpus=46 somewhat reflects the number of chromosomes. However, this many CPU cores forces your jobs to the largest servers in the cluster (with 72 CPU cores), of which there are only very few available. Given the job name (unitig to ref mapping), you could easily go for 12 or 24 CPU cores. Many more machines match that spec, which should overall result in a faster execution of your pipeline.