Unassigned clustering now creates cluster only if unitigs have high sim
Before, a long unitig was assigned a cluster, and then similarity to it was checked. Now, similarities are checked, and if two unitigs have a high similarity, only then is a new cluster created. This can work better because unitigs that are dissimilar to all other clusters and unitigs will remain NA, instead of being assigned a cluster, and then not sharing it. This may break if assemblies get to the point of many one unitig chromosomes, and those chromosomes aren't detected with contiBAIT clustering