Orientation clustering uses projection onto perp-plane instead of PC1
This is to handle the rare case when, the X and Y chromosome are particularly spread apart, and one chromosome is more fragmented than the other, the PC1 rule will fail as the more fragmented chromosome is pulled to PC1, and pushes the contiguous chromosome such that part of it crosses to the other half of PC1. Using the GLM is an imperfect fix, but accounts for more of the angle and size of the unitigs to help improve results.