Unset PEAR -t parameter
This parameters (I believe) controls the minimum length of the read before merging is attempted, and at a value of 101, was leading libraries in some samples to discard >= 75% of reads before input to bwa fastmap. However, it seems that the PEAR defaults include other reasonable constraints, so will experiment with not setting the value.