Specify default resources that scale with attempt
I would like to not specify all resources, and instead be able to specify default resources at the command line. However, specifying defaults seems to, at the moment, not work with HILBERT, and so I am using the functions to be able to scale the default resources with attempt, which is needed for the more resource intensive assemblies, where even the small rules need more resources.