Haplotype marker counting now uses dot product onto WC basis vector
Previously, used a bit of an arbitrary hack to weight libraries (weight^2) and then rescaled the counts to the original n. This process, especially with the rescaling, made it hard to interpret the variable counts. Now, the WC basis vector is projected onto the square, which allows for the WC indicator variable interpretation, and the counts are computed simply as the magnitude of the dot product onto this WC indicator variable. This allows for the haplotype marker counts to be interpreted directly as counts, whose magnitudes are now meaningful.