Marker counting now rotates WW basis vector to infer WC basis vector
Previously, the WC  basis vector was inferred by orienting the unigs, and then using a regression model to classify inverted and not-inverted unitigs, with the classification boundary lying on the WC basis vector. However, this strategy needed special case handling for one-haplotype clusters, which was brittle and could fail in surprising ways. Now, the WC basis vector is inferred by:

1. Calculating the WW basis vector by summing oriented bwa mem counts
2. Adding in inverted unitigs, and adding in "projected" unitigs, which are obtained by projecting unitigs through the WW basis vector. Adding in projected unitigs guarantees that  two haplotype clusters will be present in the data, and together with the inverted unitgs, they help to "self-supervise" the upcoming PCA by adding more points to the data plane.
3. Perform PCA to detect the data plane, for which the first two principal components form a basis.
4. Project the WW basis vector into the PCA data plane, rotate it 90 degrees to obtain the WC basis vector, and then project the WC basis vector back into library space.

This method is robust to one-haplotype clusters, are there is now no detection threshold for one-haplotype clusters that may be violated by noisy data. Additionally, by adding the projected unitigs, this method utilizes more information to "self-supervise" the PCA, and is conceptually simpler as well. It also avoids some noise that can occur with the regression model when haplotypes are imbalanced in size.